fdafazzzzzzzzzzzz

+++ development-search/src/main/java/com/sophia/web/search/SQLViewController.java
@@ -5,10 +5,10 @@ import java.util.Map;
 import javax.servlet.http.HttpServletRequest;
 import javax.validation.Valid;
 
+import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.MediaType;
 import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
 import org.springframework.ui.ModelMap;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -182,6 +182,9 @@ public class SQLViewController extends BaseController{
 	@RequestMapping(value="/delete/{code}/{id}",method=RequestMethod.POST, consumes = MediaType.APPLICATION_JSON_VALUE)
 	public Map<String, Object> deleteView(@PathVariable String code,@PathVariable String id) {
 		try {
+			if(StringUtils.isEmpty(id)){
+				throw new Exception("id不能为空");
+			}
 			sqlViewService.deleteByCode(code, id);
 			return responseOk(Constant.SUCCESS_MESSAGE);
 		} catch (Exception e) {
